Revert order of `EmbeddedDataSpecification` attributes in the schemas (#587)
Previously we mistakenly thought that attribute `dataSpecification` of
class `EmbeddedDataSpecification` was optional and we did a bugfix for
this in [aas-core-meta#328]. However, we later discovered that this was
a mistake in a graphic of the specification and `data_specification` was
indeed mandatory (see discussion in [aas-core-meta#326]). Therefore we
reverted the added `Optional` in the type hint in [aas-core-meta#329].

Sadly, the issue was not finished with this, as we forgot to also revert
the order of the attributes that had changed due to the fact that
optional attributes need to be defined behind mandatory attributes.
This had been found in #477 with regards to V3.0 and had been fixed for
v3.0 in [aas-core-meta#353]. Unfortunately we forgot to adapt this
change also for V3.1.

So now, another issue was filed in #585, notifying us of the changed
order of attributes in V3.1.

Finally, [aas-core-meta#371] reverts the order of the attributes of
`Embedded_data_specification` for V3.1 as well and we propagate the
changes to the schema files here.

[aas-core-meta#326]: https://github.com/aas-core-works/aas-core-meta/issues/326
[aas-core-meta#328]: https://github.com/aas-core-works/aas-core-meta/pull/328
[aas-core-meta#329]: https://github.com/aas-core-works/aas-core-meta/pull/329
[aas-core-meta#353]: https://github.com/aas-core-works/aas-core-meta/pull/353
[aas-core-meta#371]: https://github.com/aas-core-works/aas-core-meta/pull/371

Fixes #585